Use official Bridget release, reduce scope
