+++ dotcom-rendering/package.json
@@ -28,7 +28,7 @@
 		"@emotion/server": "11.11.0",
 		"@guardian/ab-core": "8.0.0",
 		"@guardian/braze-components": "22.2.0",
-		"@guardian/bridget": "0.0.0-2025-05-28-snapshot-2",
+		"@guardian/bridget": "8.4.0",
 		"@guardian/browserslist-config": "6.1.0",
 		"@guardian/cdk": "61.4.0",
 		"@guardian/commercial": "26.1.2",
