--- dotcom-rendering/src/components/NativePlatformWrapper.importable.tsx
@@ -1,17 +1,0 @@
-import { NativePlatform } from '@guardian/bridget';
-import { useEffect } from 'react';
-import { useNativePlatform } from '../lib/useNativePlatform';
-
-export const NativePlatformWrapper = () => {
-	const nativePlatform = useNativePlatform();
-
-	useEffect(() => {
-		if (nativePlatform !== undefined) {
-			const className =
-				nativePlatform === NativePlatform.ios ? 'ios' : 'android';
-
-			document.body.classList.add(className);
-		}
-	}, [nativePlatform]);
-	return null;
-};
